Change member_id_t to uint16_t

+++ type_definitions/primitives.hpp
@@ -4,7 +4,7 @@
 #include <cstdint>
 
 typedef int service_id_t;
-typedef int member_id_t;
+typedef uint16_t member_id_t;
 typedef uint8_t message_id_t;
 typedef CryptoPP::SecByteBlock blinded_secret_t;
 typedef CryptoPP::SecByteBlock secret_t;
